Added friends group option to student preferences  (#6)
* Changed random sampler to not coerce random floats into ints

* Added friends group option as a vector preference

* Changed str to String type

* Used String::new instead of from

+++ src/optimizers/mcmc/mcmc_naive.rs
@@ -35,8 +35,8 @@ impl MCMCOptimizer for MCMCNaive{
         // Uniform, random sampling
         let size = self.ballots.students.len();
 
-        let student = self.gen_range(0 as f64, size as f64) as usize;
-        let mut house = self.gen_range(0 as f64, (schedule.len() -1) as f64) as usize;
+        let student = self.gen_range(0, size);
+        let mut house = self.gen_range(0, schedule.len() -1);
 
         if house >= schedule.len() { // ensure we don't get the same house
             house += 1;
